refactor: useideas hook
